MM-05: Responding to code review 2
